refactor(review-station): 지하철역 관련 request/response DTO를 application 계층으로 이관합니다.
ref: https://github.com/next-step/atdd-subway-map/pull/604#discussion_r1089979133
